Slightly better waitcomment to deal with linewrapping
Script was failing with operations on either very short or very long directories since it depended on how many lines the sudo log output was. Turns out it isn't always exactly two lines. Alternative would be to disable linewrapping in sudoers but this is better as long as someone doesn't have a file named TTY.